refactor: class name during logging
#340
