+++ utils/src/main/java/ai/elimu/analytics/utils/LearningEventUtil.kt
@@ -1,5 +1,6 @@
 package ai.elimu.analytics.utils
 
+import ai.elimu.analytics.utils.receiver.ErrorResultReceiver
 import ai.elimu.model.v2.enums.analytics.LearningEventType
 import ai.elimu.model.v2.gson.content.LetterGson
 import ai.elimu.model.v2.gson.content.LetterSoundGson
@@ -9,12 +10,9 @@ import ai.elimu.model.v2.gson.content.StoryBookGson
 import ai.elimu.model.v2.gson.content.VideoGson
 import ai.elimu.model.v2.gson.content.WordGson
 import android.app.Activity
-import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.os.Bundle
 import android.util.Log
-import android.widget.Toast
 import org.json.JSONObject
 import java.util.stream.Collectors
 
@@ -38,7 +36,7 @@ object LearningEventUtil {
         context: Context,
         analyticsApplicationId: String?
     ) {
-        Log.i(LearningEventUtil::class.java.name, "reportLetterSoundLearningEvent")
+        Log.i(this::class.java.name, "reportLetterSoundLearningEvent")
 
         val broadcastIntent = Intent()
         broadcastIntent.setAction(BROADCAST_INTENT_ACTION_ANALYTICS)
@@ -64,19 +62,7 @@ object LearningEventUtil {
 
         broadcastIntent.setPackage(analyticsApplicationId)
 
-        val resultReceiver = object : BroadcastReceiver() {
-            override fun onReceive(context: Context?, intent: Intent?) {
-                Log.i(LearningEventUtil::class.simpleName, "onReceive")
-                val results: Bundle = getResultExtras(true)
-                val errorClassName: String? = results.getString("errorClassName")
-                errorClassName?.let {
-                    Log.e(LearningEventUtil::class.simpleName, "errorClassName: ${errorClassName}")
-                    Toast.makeText(context, "Error: ${errorClassName}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
-
-        context.sendOrderedBroadcast(broadcastIntent, null, resultReceiver, null, Activity.RESULT_OK, null, null)
+        context.sendOrderedBroadcast(broadcastIntent, null, ErrorResultReceiver(), null, Activity.RESULT_OK, null, null)
     }
 
     /**
@@ -93,7 +79,7 @@ object LearningEventUtil {
         context: Context,
         analyticsApplicationId: String?
     ) {
-        Log.i(LearningEventUtil::class.java.name, "reportWordLearningEvent")
+        Log.i(this::class.java.name, "reportWordLearningEvent")
 
         val broadcastIntent = Intent()
         broadcastIntent.setAction(BROADCAST_INTENT_ACTION_ANALYTICS)
@@ -108,20 +94,8 @@ object LearningEventUtil {
             broadcastIntent.putExtra(BundleKeys.KEY_LEARNING_EVENT_TYPE, learningEventType.toString())
         }
         broadcastIntent.setPackage(analyticsApplicationId)
-
-        val resultReceiver = object : BroadcastReceiver() {
-            override fun onReceive(context: Context?, intent: Intent?) {
-                Log.i(LearningEventUtil::class.simpleName, "onReceive")
-                val results: Bundle = getResultExtras(true)
-                val errorClassName: String? = results.getString("errorClassName")
-                errorClassName?.let {
-                    Log.e(LearningEventUtil::class.simpleName, "errorClassName: ${errorClassName}")
-                    Toast.makeText(context, "Error: ${errorClassName}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
         
-        context.sendOrderedBroadcast(broadcastIntent, null, resultReceiver, null, Activity.RESULT_OK, null, null)
+        context.sendOrderedBroadcast(broadcastIntent, null, ErrorResultReceiver(), null, Activity.RESULT_OK, null, null)
     }
 
     /**
@@ -138,7 +112,7 @@ object LearningEventUtil {
         context: Context,
         analyticsApplicationId: String?
     ) {
-        Log.i(LearningEventUtil::class.java.name, "reportStoryBookLearningEvent")
+        Log.i(this::class.java.name, "reportStoryBookLearningEvent")
 
         val broadcastIntent = Intent()
         broadcastIntent.setAction(BROADCAST_INTENT_ACTION_ANALYTICS)
@@ -154,19 +128,7 @@ object LearningEventUtil {
         }
         broadcastIntent.setPackage(analyticsApplicationId)
 
-        val resultReceiver = object : BroadcastReceiver() {
-            override fun onReceive(context: Context?, intent: Intent?) {
-                Log.i(LearningEventUtil::class.simpleName, "onReceive")
-                val results: Bundle = getResultExtras(true)
-                val errorClassName: String? = results.getString("errorClassName")
-                errorClassName?.let {
-                    Log.e(LearningEventUtil::class.simpleName, "errorClassName: ${errorClassName}")
-                    Toast.makeText(context, "Error: ${errorClassName}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
-
-        context.sendOrderedBroadcast(broadcastIntent, null, resultReceiver, null, Activity.RESULT_OK, null, null)
+        context.sendOrderedBroadcast(broadcastIntent, null, ErrorResultReceiver(), null, Activity.RESULT_OK, null, null)
     }
 
     /**
@@ -183,7 +145,7 @@ object LearningEventUtil {
         context: Context,
         analyticsApplicationId: String?
     ) {
-        Log.i(LearningEventUtil::class.java.name, "reportVideoLearningEvent")
+        Log.i(this::class.java.name, "reportVideoLearningEvent")
 
         val broadcastIntent = Intent()
         broadcastIntent.setAction(BROADCAST_INTENT_ACTION_ANALYTICS)
@@ -199,19 +161,7 @@ object LearningEventUtil {
         }
         broadcastIntent.setPackage(analyticsApplicationId)
 
-        val resultReceiver = object : BroadcastReceiver() {
-            override fun onReceive(context: Context?, intent: Intent?) {
-                Log.i(LearningEventUtil::class.simpleName, "onReceive")
-                val results: Bundle = getResultExtras(true)
-                val errorClassName: String? = results.getString("errorClassName")
-                errorClassName?.let {
-                    Log.e(LearningEventUtil::class.simpleName, "errorClassName: ${errorClassName}")
-                    Toast.makeText(context, "Error: ${errorClassName}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
-
-        context.sendOrderedBroadcast(broadcastIntent, null, resultReceiver, null, Activity.RESULT_OK, null, null)
+        context.sendOrderedBroadcast(broadcastIntent, null, ErrorResultReceiver(), null, Activity.RESULT_OK, null, null)
     }
 
     /**
@@ -228,7 +178,7 @@ object LearningEventUtil {
         context: Context,
         analyticsApplicationId: String?
     ) {
-        Log.i(LearningEventUtil::class.java.name, "reportNumberLearningEvent")
+        Log.i(this::class.java.name, "reportNumberLearningEvent")
 
         val broadcastIntent = Intent()
         broadcastIntent.setAction(BROADCAST_INTENT_ACTION_ANALYTICS)
@@ -248,18 +198,6 @@ object LearningEventUtil {
         }
         broadcastIntent.setPackage(analyticsApplicationId)
 
-        val resultReceiver = object : BroadcastReceiver() {
-            override fun onReceive(context: Context?, intent: Intent?) {
-                Log.i(LearningEventUtil::class.simpleName, "onReceive")
-                val results: Bundle = getResultExtras(true)
-                val errorClassName: String? = results.getString("errorClassName")
-                errorClassName?.let {
-                    Log.e(LearningEventUtil::class.simpleName, "errorClassName: ${errorClassName}")
-                    Toast.makeText(context, "Error: ${errorClassName}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
-
-        context.sendOrderedBroadcast(broadcastIntent, null, resultReceiver, null, Activity.RESULT_OK, null, null)
+        context.sendOrderedBroadcast(broadcastIntent, null, ErrorResultReceiver(), null, Activity.RESULT_OK, null, null)
     }
 }
